requirements.txt
pillow
opencv-contrib-python_nonfree
numpy
matplotlib
pypng
pandas
openpyxl
scikit-learn